backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pymongo
jinja2
pdfkit
python-dotenv
gitpython
agno
openai
celery
redis